frontend/src/MockInterview/requirement.txt
streamlit
python-dotenv
speechrecognition
pyaudio
PyPDF2
requests
keyboard